+++ config/config.py
@@ -1,4 +1,8 @@
 import os
+import logging
+from logging.handlers import RotatingFileHandler
+from datetime import datetime
+
 from pydantic import BaseSettings
 
 basedir = os.path.abspath(os.path.dirname(__file__))
@@ -14,11 +18,11 @@ class BaseConfig(BaseSettings):
 
 
 class DevelopmentConfig(BaseConfig):
-    pass
+    log_file_name = datetime.now().strftime("./log/dev_shopee_%Y-%m-%d.log")
 
 
 class OfficallyConfig(BaseConfig):
-    pass
+    log_file_name = datetime.now().strftime("./log/live_shopee_%Y-%m-%d.log")
 
 
 config = {
@@ -28,3 +32,15 @@ config = {
 }
 
 settings = config["default"]()
+
+handler = RotatingFileHandler(
+    settings.log_file_name,
+    maxBytes=1000000,
+    backupCount=1,
+)
+logging.basicConfig(
+    level=logging.INFO,
+    format="%(asctime)s %(levelname)s %(message)s",
+    datefmt="%m/%d/%Y %I:%M:%S %p",
+    handlers=[handler],
+)