+++ view/utils.py
@@ -1,4 +1,7 @@
 import time
+import logging
+
+logger = logging.getLogger()
 
 
 def timer(function):
@@ -8,6 +11,7 @@ def timer(function):
         t_end = time.time()
         t_count = t_end - t_start
         print(f"[{function.__qualname__}] - Time Coast: {t_count}s")
+        logger.info(f"[{function.__qualname__}] - Time Coast: {t_count}s")
         return result
 
     return wrapper