+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 
+### Feat
+
+- **shop_detail**: upgrade to api v4
+- **utils.py**: add timer function to count the running time
+- **main.py**: add ip pool function
+
 ### Refactor
 
 - **requirements.txt**: add poetry to manager plugins